fixup: tweak to get Flow happy
