+++ app/src/config/constants.js
@@ -2,6 +2,8 @@
 
 import type { DevInternalFlag } from './types'
 
+export const CONFIG_VERSION_LATEST: 1 = 1
+
 export const DEV_INTERNAL_FLAGS: Array<DevInternalFlag> = [
   'allPipetteConfig',
   'enableBundleUpload',
