+++ app/src/config/schema-types.js
@@ -1,8 +1,25 @@
 // @flow
 import type { LogLevel } from '../logger'
 
+import type { LogLevel } from '../logger'
+
+export type UrlProtocol = 'file:' | 'http:'
+
+export type UpdateChannel = 'latest' | 'beta' | 'alpha'
+
+export type DiscoveryCandidates = string | Array<string>
+
+export type DevInternalFlag =
+  | 'allPipetteConfig'
+  | 'enableBundleUpload'
+  | 'enableRobotCalCheck'
+
+export type FeatureFlags = $Shape<{|
+  [DevInternalFlag]: boolean | void,
+|}>
+
 export type ConfigV0 = $ReadOnly<{|
-  configFileVersion: 0,
+  version: 0,
   devtools: boolean,
   reinstallDevtools: boolean,
 
@@ -43,10 +60,11 @@ export type ConfigV0 = $ReadOnly<{|
     seenOptIn: boolean,
   |}>,
 
-  // deprecated; remove with first migration
-  p10WarningSeen: $ReadOnly<{|
+  // deprecated
+  p10WarningSeen: $ReadOnly<{
     [id: string]: ?boolean,
-  |}>,
+    ...,
+  }>,
 
   support: $ReadOnly<{|
     userId: string,
@@ -71,19 +89,13 @@ export type ConfigV0 = $ReadOnly<{|
   devInternal?: $ReadOnly<FeatureFlags>,
 |}>
 
-// When a V0 config is compared to V1's defaults the first time,
-// the version number field will be added by default (even without migration)
-// because of the way electron-store handles missing fields.
-export type ConfigIntermediateV0ToV1 = $ReadOnly<{|
-  ...ConfigV0,
-  version: 0,
-|}>
-
 export type ConfigV1 = $ReadOnly<{|
   ...ConfigV0,
   version: 1,
   discovery: $ReadOnly<{|
     candidates: DiscoveryCandidates,
-    disableDiscoveryCache: boolean,
+    disableCache: boolean,
   |}>,
 |}>
+
+export type Config = ConfigV1
