+++ app/src/config/types.js
@@ -1,24 +1,5 @@
 // @flow
-import type {
-  ConfigV0,
-  ConfigV1,
-  ConfigIntermediateV0ToV1,
-} from './ConfigTypes'
-
-export type UrlProtocol = 'file:' | 'http:'
-
-export type UpdateChannel = 'latest' | 'beta' | 'alpha'
-
-export type DiscoveryCandidates = string | Array<string>
-
-export type DevInternalFlag =
-  | 'allPipetteConfig'
-  | 'enableBundleUpload'
-  | 'enableRobotCalCheck'
-
-export type FeatureFlags = $Shape<{|
-  [DevInternalFlag]: boolean | void,
-|}>
+export type * from './schema-types'
 
 export type UpdateConfigAction = {|
   type: 'config:UPDATE',
@@ -39,8 +20,6 @@ export type SetConfigAction = {|
     value: any,
   |},
 |}
-
-export type Config = ConfigV0 | ConfigV1 | ConfigIntermediateV0ToV1
 
 export type ConfigAction =
   | UpdateConfigAction
